requirements.txt
asgiref==3.5.2
Django==4.0.6
django-modeltranslation==0.18.2
gunicorn==20.1.0
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
